pyproject.toml
[tool.poetry]
name = "app_campze"
version = "0.1.0"
description = "Camp Management System"
authors = ["Jack Mthembu"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"  # Using Python 3.11 for better compatibility
Flask = "2.2.5"
Flask-SQLAlchemy = "3.0.2"
Flask-Login = "0.6.2"
Flask-WTF = "1.1.1"
Flask-Mail = "0.9.1"
Flask-Migrate = "4.0.4"
Flask-CORS = "4.0.0"
SQLAlchemy = "2.0.19"
Werkzeug = "2.2.3"
WTForms = "3.0.1"
python-dotenv = "1.0.0"
email-validator = "2.0.0"
Pillow = "10.0.0"
gunicorn = "21.2.0"
psycopg2-binary = "2.9.9"
alembic = "1.12.0"
bcrypt = "4.0.1"
requests = "2.31.0"
stripe = "7.0.0"
redis = "5.0.1"
celery = "5.3.4"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.3"
pytest-cov = "4.1.0"
black = "23.11.0"
flake8 = "6.1.0"
mypy = "1.7.1"
isort = "5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 